Initial commit: There is an index file with all of git commands. Fixed rust structure. Get all modified files in repo. Get a file properties to send to the commit page.
Init is a WIP.
Clone is a WIP.
Commit is a WIP.
Revision graph is a WIP.

TODO:
Revision Graph:
Need to fix the data that goes to the js.
Added context menu per node.
Now the code is working on real repo. Still need to fix the edges.
Log:
There is no View yet but the logic is a WIP.
Working but still no good relation.